+++ abi/src/main/scala/scalether/abi/tuple/Tuple2Type.scala
@@ -1,5 +1,7 @@
 package scalether.abi.tuple
 
+import java.math.BigInteger
+
 import scalether.abi.{Decoded, Type, Uint256Type}
 
 import scala.collection.mutable.ListBuffer
@@ -13,13 +15,13 @@ class Tuple2Type[T1, T2](val type1: Type[T1], val type2: Type[T2]) extends Tuple
     val head = ListBuffer[Byte]()
     val tail = ListBuffer[Byte]()
     if (type1.dynamic) {
-      head ++= Uint256Type.encode(headSize + tail.size)
+      head ++= Uint256Type.encode(BigInteger.valueOf(headSize + tail.size))
       tail ++= type1.encode(value._1)
     } else {
       head ++= type1.encode(value._1)
     } 
     if (type2.dynamic) {
-      head ++= Uint256Type.encode(headSize + tail.size)
+      head ++= Uint256Type.encode(BigInteger.valueOf(headSize + tail.size))
       tail ++= type2.encode(value._2)
     } else {
       head ++= type2.encode(value._2)
