+++ core/src/main/scala/scalether/core/transaction/SimpleTransactionSender.scala
@@ -1,12 +1,14 @@
 package scalether.core.transaction
 
+import java.math.BigInteger
+
 import cats.MonadError
 import scalether.core.{Ethereum, TransactionSender}
 import scalether.core.request.Transaction
 
 import scala.language.higherKinds
 
-class SimpleTransactionSender[F[_]](ethereum: Ethereum[F], from: String, gas: BigInt, gasPrice: BigInt)(implicit me: MonadError[F, Throwable])
+class SimpleTransactionSender[F[_]](ethereum: Ethereum[F], from: String, gas: BigInteger, gasPrice: BigInteger)(implicit me: MonadError[F, Throwable])
   extends TransactionSender[F] {
 
   def call(transaction: Transaction) =
